Image similar search/requirements.txt
acres==0.5.0
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
build==1.4.2
certifi==2026.2.25
charset-normalizer==3.4.6
chromadb==1.5.5
ci-info==0.4.0
click==8.3.1
configobj==5.0.9
configparser==7.2.0
durationpy==0.10
etelemetry==0.3.1
fastapi==0.135.2
filelock==3.25.2
flatbuffers==25.12.19
frontend==0.0.3
fsspec==2026.2.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
looseversion==1.3.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
networkx==3.6.1
nibabel==5.4.2
nipype==1.11.0
numpy==2.4.3
oauthlib==3.3.1
onnxruntime==1.24.4
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==3.0.1
pathlib==1.0.1
pillow==12.1.1
protobuf==6.33.6
prov==2.1.1
puremagic==2.1.1
pybase64==1.4.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.27.2.2
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pyxnat==1.6.4
PyYAML==6.0.3
rdflib==7.6.0
referencing==0.37.0
requests==2.33.0
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
scipy==1.17.1
setuptools==81.0.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
traits==7.1.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
zipp==3.23.0
